+++ RocketApp.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -60,7 +60,7 @@
       "location" : "https://github.com/Qase/swift-core",
       "state" : {
         "branch" : "develop",
-        "revision" : "d5ba77e15c2f633f62861ec2b93bf98df176cef7"
+        "revision" : "fb5587c91207785baf03fee9b89a10671d5148c3"
       }
     },
     {
