+++ RocketApp/Domain/Sources/RocketsClient/RocketsClient+liveAsync.swift
@@ -7,31 +7,6 @@ import Networking
 import RequestBuilder
 
 public extension RocketsClient {
-  static var liveReact: Self {
-    @Dependency(\.networkClientType) var networkClientType
-    @Dependency(\.rocketConverter) var rocketConverter
-    @Dependency(\.rocketsConverter) var rocketsConverter
-
-    return Self(
-      getRocket: { id in
-        try await Request(endpoint: URLs.baseURL + "/v3/rockets/\(id)")
-          .execute(using: networkClientType)
-          .mapErrorReporting(to: RocketsClientError.networkError)
-          .convertToDomainModel(using: rocketConverter)
-          .eraseToAnyPublisher()
-          .async()
-      },
-      getAllRockets: {
-        try await Request(endpoint: URLs.baseURL + "/v3/rockets")
-          .execute(using: networkClientType)
-          .mapErrorReporting(to: RocketsClientError.networkError)
-          .convertToDomainModel(using: rocketsConverter)
-          .eraseToAnyPublisher()
-          .async()
-      }
-    )
-  }
-
   static var liveAsync: Self {
     @Dependency(\.networkClientType) var networkClientType
     @Dependency(\.rocketConverter) var rocketConverter
