+++ RocketApp/Domain/Sources/RocketsClient/RocketsErrorManager/RocketsClientAsyncError.swift
@@ -13,7 +13,7 @@ public enum RocketsClientAsyncError: Error, Equatable, CustomDebugStringConverti
   public var description: String {
     switch self {
     case let .networkError(error):
-      return "network error \(error)"
+      return "network error \(error.cause.description)"
     case .modelConversionError:
       return "model conversion error"
     case .undefinedError:
